refactor: 나의 모임 목록 조회 로직 GatheringJoinService로 이동
