+++ momoim-external-api/src/main/java/com/triplem/momoim/api/gathering/controller/GatheringController.java
@@ -4,7 +4,6 @@ import com.triplem.momoim.api.common.ApiResponse;
 import com.triplem.momoim.api.gathering.dto.GatheringCategoryInformation;
 import com.triplem.momoim.api.gathering.service.GatheringService;
 import com.triplem.momoim.auth.utils.SecurityUtil;
-import com.triplem.momoim.core.common.PaginationInformation;
 import com.triplem.momoim.core.common.SortOrder;
 import com.triplem.momoim.core.domain.gathering.dto.GatheringDetail;
 import com.triplem.momoim.core.domain.gathering.dto.GatheringPreview;
@@ -21,11 +20,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -54,23 +51,6 @@ public class GatheringController {
         @Parameter(hidden = true) GatheringSearchOption option) {
         List<GatheringPreview> gatherings = gatheringService.searchGatherings(option);
         return ApiResponse.success(gatherings);
-    }
-
-    @GetMapping("/joined")
-    @PreAuthorize("hasAnyRole('ROLE_USER')")
-    @ApiResponses(value = {
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "요청 성공")
-    })
-    @Operation(operationId = "내가 속한 모임 조회", summary = "내가 속한 모임 조회", tags = {"gatherings"}, description = "내가 속한 모임 조회")
-    public ApiResponse<List<GatheringPreview>> getMyGatherings(
-        @Parameter(description = "페이징 offset") @RequestParam int offset,
-        @Parameter(description = "페이징 limit") @RequestParam int limit) {
-        Long userId = SecurityUtil.getMemberIdByPrincipal();
-        List<GatheringPreview> myGatherings = gatheringService.getMyGatherings(
-            userId,
-            new PaginationInformation(offset, limit)
-        );
-        return ApiResponse.success(myGatherings);
     }
 
     @GetMapping("/categories")
